requirements.txt
anthropic>=0.40.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
ta>=0.11.0
python-telegram-bot>=20.0
playwright>=1.40.0
apscheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
feedparser>=6.0.0
scipy>=1.11.0
lxml>=4.9.0
